requirements.txt
autopep8==1.4.4
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
Click==7.0
dnspython==2.2.1
email-validator==1.2.1
Flask==1.1.1
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.9.3
pycodestyle==2.5.0
pycparser==2.21
requests==2.28.1
SQLAlchemy==1.4.36
urllib3==1.26.10
Werkzeug==0.16.0
WTForms==3.0.1
